# perf: set recursive to false since always dealing with depth 2