--- a/R/brackets.R
+++ b/R/brackets.R
@@ -83,7 +83,7 @@
   evals <- tf_evaluate(x, arg = j)
   if (!interpolate) {
     new_j <- map2(j, ensure_list(tf_arg(x)), \(x, y) !(x %in% y))
-    if (any(unlist(new_j, use.names = FALSE))) {
+    if (any(unlist(new_j, recursive = FALSE, use.names = FALSE))) {
       cli::cli_warn(c(
         i = "{.code interpolate = FALSE} & no values present for some {.arg j}",
         x = "{.code NA}s created."
@@ -95,7 +95,7 @@
 
   if (matrix) {
     ret <- do.call(rbind, evals)
-    j <- unlist(j, use.names = FALSE)
+    j <- unlist(j, recursive = FALSE, use.names = FALSE)
     colnames(ret) <- j
     rownames(ret) <- names(x)
     return(structure(ret, arg = j))