Since arg and evaluations are always max depth of two, we can set recursive to false.
Minor optimization, but a free-lunch:

``` r
x <- tf::tf_rgp(1e5, 100) |> tf::tf_sparsify() |> tf::tf_arg()
bench::mark(
  unlist(x, use.names = FALSE),
  unlist(x, recursive = FALSE, use.names = FALSE)
)
#> # A tibble: 2 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                          <bch:> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 unlist(x, use.names = FALSE)        14.9ms 15.1ms      66.2    38.1MB    99.2 
#> 2 unlist(x, recursive = FALSE, use.n… 13.8ms 14.7ms      68.3    38.1MB     4.55
```

<sup>Created on 2024-12-07 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>