@fabian-s 2024-12-09
> arg and evaluations are always max depth of two

won't (necessarily) be like that anymore once we implement multivariate functions (or fields), so not worth the small performance gain I think.